% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline}
\alias{baseline}
\title{Baseline data table}
\format{
A data frame with seven variables: \code{pt_id}, \code{fname},
\code{sname}, \code{dob}, \code{rec_date}, \code{wd_stat} and \code{wd_date}
}
\usage{
baseline
}
\description{
Information of patients
}
\keyword{datasets}
