\name{cauchyrnd}
\alias{cauchyrnd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Cauchy Destribution Random Numbers}
\description{
  Calculate random numbers from Cauchy distribution.
}
\usage{
cauchyrnd(ml, md=1, m, n=1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ml}{ meassury of location }
  \item{md}{ measury of dispersion }
  \item{m}{ a vector with random number or the number of rows }
  \item{n}{ the number of columns for the matrix, optional n=1 }
}
\details{
  Use cauchyinv.
}  
\author{ Zornitsa Manolova }

\seealso{ objects to See Also as \code{\link{cauchyinv}}}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ cauchy }
\keyword{ random }% __ONLY ONE__ keyword per line
