\name{chi2modus}
\alias{chi2modus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Chisquare Modus }
\description{
 Calculate the Modus of Chisquare destribution by given degree of freedom.
}
\usage{
chi2modus(dof)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dof}{ degree of freedom }
}

\value{
  \item{ }{Return the modus of chi square distribution = max(chi2pdf(DF))
}

}
\author{ Zornitsa Manolova }


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ Chisquare }
\keyword{ Modus }
