\name{dbt.Statistics-package}
\alias{dbt.Statistics-package}
\alias{dbt.Statistics}
\docType{package}
\title{Statistics
}
\description{
A lot of statistical tests
}
\details{   
\tabular{ll}{
Package: \tab dbt.Statistics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-30\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{Zornitsa Manolova}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
