\name{dbt_summary}
\alias{dbt_summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Summary }
\description{
  Descriptive statistics for the features 
}
\usage{
dbt_summary(Data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ numeric matrix, cases in rows, features in colums }
}
\value{
  \item{ elements }{ (meassure of location) array containing the statistics as described in column names}
}
\author{ Zornitsa Manolova }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ summary }
