% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{discharged}
\alias{discharged}
\title{\enumerate{
\item Return number of discharged patient
}}
\usage{
discharged()
}
\value{
A string with number of discharged patient
}
\description{
\enumerate{
\item Return number of discharged patient
}
}
\examples{
library(dplyr)
data("inpatient_main")
discharged()
}
