\name{gmean}
\alias{gmean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Geometrical Mean}
\description{
Calcultes geometrical mean for a given vector of data
}
\usage{
gmean(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ numeric vector or matrix }
}
\details{
  The issue ist always a single value. Matrices are not speacially treated
  to calculate the value column by column. 
}
\value{
  \item{m}{ geometrical mean}

}
\author{Zornitsa Manolova}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ mean }
