\name{GrubbsOutlierTest}
\alias{GrubbsOutlierTest}
\title{ performs an Grubbs test for identifying outliers  returns the possible Outliers and the corresponding Pvalues }
\usage{GrubbsOutlierTest(X)}
\arguments{
\item{X}{(1:n) data,  may contain NaN}
}
\value{Outlier(1:o):            possible outliers in X}
\description{
 performs an Grubbs test for identifying outliers  returns the possible Outliers and the corresponding Pvalues }
\author{
MT 08/2015 nach matlab: ALU Aug. 2015
}
