\name{hmean}
\alias{hmean}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Harmonic Mean }
\description{
Calcultes harmonic mean for a given vector of data
}
\usage{
hmean(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a numeric vector}
}
\author{ Zornitsa Manolova }


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ mean }
