\name{HyperGeometricCdf}
\alias{HyperGeometricCdf}
\title{ hypergeometric_cdf evaluates the Hypergeometric CDF. }
\arguments{
\item{xIn[1:i]}{the argument of the CDF                                MATLAB,hygecdf(X,M,K,N): X}
\item{nIn[1:i]}{number of balls selected 0 <= n <= l                   MATLAB,hygecdf(X,M,K,N): N}
\item{mIn[1:i]}{number of white balls in the population:  0 <= m <= l  MATLAB,hygecdf(X,M,K,N): K}
\item{lIn[1:i]}{the number of balls to select from  0 <= l             MATLAB,hygecdf(X,M,K,N): M i>=1}
}
\value{cdf          the value of the Hypergeometric CDF at x}
\description{
 hypergeometric_cdf evaluates the Hypergeometric CDF. }
\author{
John Burkardt}
