\name{informationRetrieval}
\alias{informationRetrieval}
\title{ Bei gegebener Referenzliste Full und Vergleichliste Curt werden die Variablen des Information Retrievals berechnet }
\arguments{
\item{vCurt[1:m]}{vector,}
\item{vFull[1:n]}{vector,}
}
\value{Precision           vector}
\description{
 Bei gegebener Referenzliste Full und Vergleichliste Curt werden die Variablen des Information Retrievals berechnet }
\author{
Michael Thrun
}
