% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inpatient_2018}
\alias{inpatient_2018}
\title{Inpatient 2018 table}
\format{
A data frame with four variables: \code{pt_id}, \code{record_id},
\code{admidate}, \code{disdate} and \code{batch}
}
\usage{
inpatient_2018
}
\description{
Information of patients with admision or descharge in 2018
}
\keyword{datasets}
