% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inpatient_diag}
\alias{inpatient_diag}
\title{Inpatient diag data table}
\format{
A data frame with four variables: \code{record_id}, \code{arr_index},
\code{level} and \code{diag_icd10}
}
\usage{
inpatient_diag
}
\description{
Information of patients diagnosis
}
\keyword{datasets}
