% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inpatient_main}
\alias{inpatient_main}
\title{Inpatient main data table}
\format{
A data frame with five variables: \code{pt_id}, \code{record_id},
\code{admidate}, \code{disdate} and \code{batch}
}
\usage{
inpatient_main
}
\description{
Information of patients
}
\keyword{datasets}
