\name{IR.AUCandFbetha}
\alias{IR.AUCandFbetha}
\title{ Berechnet gaengige Variablen des Information Retriveal }
\arguments{
\item{Precision[1:m]}{vector, positive predictive value) is the fraction of retrieved instances that are relevant}
\item{Recall[1:n]}{vector, is the fraction of relevant instances that are retrieved, known as sensitivity}
\item{betha}{paramter for Fbetha, default=0.5}
}
\value{AUC                 vector, AREA of Recall and Precision}
\description{
 Berechnet gaengige Variablen des Information Retriveal }
\author{
Michael Thrun
}
