\name{LogBinomialCoeff}
\alias{LogBinomialCoeff}
\title{ LogBinomialCoeff computes the (natural) logarithm of the Binomial coefficient. LogBino = log( binomialcoeff(N,K) ) =log(n!/( k! * (n-k)!)); }
\arguments{
\item{nIn, kIn}{for binoialcoeff n over k both can  be vectors}
}
\value{LogBino  the logarithm of binoialcoeff(N,K).}
\description{
 LogBinomialCoeff computes the (natural) logarithm of the Binomial coefficient. LogBino = log( binomialcoeff(N,K) ) =log(n!/( k! * (n-k)!)); }
\author{
John Burkardt}
