\name{LogFactorial}
\alias{LogFactorial}
\title{ LogFactorial returns the (natural) logarithm of N!. with   N! = Product ( 1 <= I <= N ) }
\arguments{
\item{nIn[1:k]}{input to facultaet}
}
\value{LogFac[1:k]  the logarithm of N! LogFac = log(n!)}
\description{
 LogFactorial returns the (natural) logarithm of N!. with   N! = Product ( 1 <= I <= N ) }
\author{
John Burkardt}
