\name{maxrobust}
\alias{maxrobust}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Robust Empirical Maximum }
\description{
  Maxhat is the mean between max(x) and the top percentile that is different form max(x).

}
\usage{
maxrobust(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a numeric vector  }
}

\author{ Zornitsa Manolova }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ maximum }
\keyword{ max }
