\name{modus}
\alias{modus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Mode }
\description{
  Compute the sample mode.
}
\usage{
modus(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{   numeric vector containing the values whose mode is to be computed }
}

\value{          
  \item{ modus }{    the most frequent value in x}
  \item{ mnumber }{   frequency of appearance of mode in x }
  \item{elements}{    the different elements of x, in decreasing order of the frequency of their occurance}
  \item{enumber}{   frequency of appearance of elements in x} 
}
\author{ Zornitsa Manolova }


\keyword{ mode }
