\name{nanamad}
\alias{nanamad}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Nanamad }
\description{
  Calculate median of absolute differences from median.
}
\usage{
nanamad(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ numeric vector or matrix }
}
\details{
 Use nanmad(x)
}
\author{ Zornitsa Manolova }

\seealso{  \code{\link{nanmad}}, ~~~ }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ mad }
\keyword{ amad }% __ONLY ONE__ keyword per line
