\name{nanmedian}
\alias{nanmedian}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ nanmedian  }
\description{
     Calculate median of Data.
}
\usage{
nanmedian(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ numeric vector or matrix }
}
\author{Michael Thrun}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{median}
\keyword{ nanmedian }% __ONLY ONE__ keyword per line
