\name{normrnd}
\alias{normrnd}
\title{ generates random numbers from normal distribution with mean parameter mu and standart deviation parameter sigma }
\arguments{
\item{mu}{mean parameter}
\item{sigma}{standart deviation parameter}
\item{n}{number of rows of OUTPUT matrix}
\item{m}{number of cols of OUTPUT matrix}
}
\value{NormMat	matrix(n,m) filled with random numbers from normal distribution}
\description{
 generates random numbers from normal distribution with mean parameter mu and standart deviation parameter sigma }
\author{
RG
}
