\name{percentiles}
\alias{percentiles}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Percentiles }
\description{
  Calculate the percentile of x
}
\usage{
percentiles(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ numeric vectors whose sample percentile are wanted. Na and NaN values are ignored. }
}

\value{
  \item{ p }{  vector of percentiles}         
}
\author{ Zornitsa Manolova }

\examples{
  
}
\keyword{ percentile }
