% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{plot_freqs}
\alias{plot_freqs}
\title{\enumerate{
\item Barplot frequencies
}}
\usage{
plot_freqs()
}
\value{
A barplot of frequencies of patient hospitalisations days
}
\description{
\enumerate{
\item Barplot frequencies
}
}
\examples{
library(dplyr)
data("baseline")
data("inpatient_main")
data("inpatient_diag")
data("inpatient_2018")
plot_freqs()
}
