\name{IR.PrecisionRecall}
\alias{IR.PrecisionRecall}
\title{ Berechnet gaengige Variablen des Information Retriveal }
\arguments{
\item{retrieved[1:m]}{vector or list, set of retrieved  (e.g. the list of objects produced)}
\item{relevant[1:n]}{vector of all objects that are relevant for a certain topic}

\item{betha}{optional, paramter for Fbetha, default=0.5}
}
\value{Precision            the fraction of retrieved objects that are relevant to the find}
\description{
 Berechnet gaengige Variablen des Information Retriveal }
\author{
Michael Thrun
}
