\name{rankdata}
\alias{rankdata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Rank Data }
\description{
  Calculate the ranks of data in each column and sort it ascending. 
}
\usage{
rankdata(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a numeric matrix }
}
\value{
  a matrix with the ranks column by column
}
\author{ Zornitsa Manolova }

\seealso{  \code{\link{ranks}} }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ rank }
