\name{rankstied}
\alias{rankstied}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Adjusted Ranks }
\description{
  Calculate ranks adjusted for ties
}
\usage{
rankstied(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ numeric vector }
}
\details{
  The Algorithm come from octave
}
\value{
    \item{ y }{   vector of ranks}
}
\author{ Zornitsa Manolova }

\keyword{ rank }

