% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{return_first_asthma}
\alias{return_first_asthma}
\title{first asthma record}
\usage{
return_first_asthma()
}
\value{
A tibble with the first hospitalisation record because of J45
}
\description{
first asthma record
}
\examples{
library(dplyr)
library(stringr)
data("inpatient_diag")
data("inpatient_main")
df <- return_first_asthma
}
