% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{return_first_asthma_2}
\alias{return_first_asthma_2}
\title{\enumerate{
\item Types of diagnostics
}}
\usage{
return_first_asthma_2()
}
\value{
Create a dataset called first_asthma_2
}
\description{
\enumerate{
\item Types of diagnostics
}
}
\examples{
library(dplyr)
library(stringr)
data("baseline")
data("inpatient_main")
data("inpatient_diag")
df <- return_first_asthma_2()
}
