% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{return_inpatient_2018_dur}
\alias{return_inpatient_2018_dur}
\title{\enumerate{
\item Duration of each admission.
}}
\usage{
return_inpatient_2018_dur()
}
\value{
A tibble with a duation of hospitalisations
}
\description{
\enumerate{
\item Duration of each admission.
}
}
\examples{
library(dplyr)
data("baseline")
data("inpatient_2018")
return_inpatient_2018_dur()
}
