% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{return_smith_list}
\alias{return_smith_list}
\title{\enumerate{
\item Return smith list
}}
\usage{
return_smith_list()
}
\value{
A tibble data
}
\description{
\enumerate{
\item Return smith list
}
}
\examples{
library(dplyr)
library(stringr)
data("baseline")
df <- return_smith_list()
}
