\name{statistics-internal}

\alias{unique_labelled}
\alias{sort_labelled}
\alias{names_values}
\alias{to_labels}
\alias{getName_statistics}
\alias{missing_values}
\alias{missing_values<-}

\title{Internal functions}

\description{
  Functions for internal use only.
}

\keyword{internal}
