\name{About the statistics package}

\alias{statistics}

\docType{package}

\title{
Utility functions for an intro stats class.
}

\description{
Plot tail or interval areas for the normal or t distributions, and draw binomial
distributions with highlighted outcome probabilities. Summarise discrete random
variables via their mean and standard deviation. Compute pooled standard deviations
and Welch-Satterthwaite degrees of freedom, then run teaching-friendly tests such as
a Fligner-Killeen variance check plus one-way ANOVA, and t tests that pair an
Ansari-Bradley variance check with both pooled and Welch variants.
}

\seealso{dnorm, pnorm, dbinom}

\details{
    \tabular{ll}{
    Package: \tab statistics\cr
    Type: \tab Package\cr
    Version: \tab 0.12\cr
    Date: \tab 2026-03-31\cr
    License: \tab GPL-v3\cr
    }
}
\author{
Adrian Dusa

Maintainer: Adrian Dusa (dusa.adrian@unibuc.ro)
}

\keyword{misc}
