\name{symlognpdf}
\alias{symlognpdf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ symlognpdf }
\description{
       
 for M>0 same as lognpdf(Data,M,S);    
 for M < 0: mirrored at y axis .      

}
\usage{
symlognpdf(Data, M, S)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Data}{ a numerical vector }
  \item{M}{ mean }
  \item{S}{ standard deviation of log normal }
}

\author{ Zornitsa Manolova }


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
