\name{symlognSigmaMue}
\alias{symlognSigmaMue}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ SymlognSigmaMue }
\description{
  Calculate the sigma and mue parameter for the lognormal distribution.
}
\usage{
symlognSigmaMue(M, S)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{M}{ mean value }
  \item{S}{ standard deviation value }
}
\value{
  \item{sig }{sigma parameter from the normal distribution}
  \item{mu }{mue parameter from the normal distribution}
 }

\author{ Zornitsa Manolova }


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ lognormal }
\keyword{ distribution }
