\name{unidrnd}
\alias{unidrnd}
\title{ unidrnd generates random numbers for the discrete uniform distribution with maximum N.}
\usage{unidrnd(N,m,n)}
\arguments{
  \item{N}{Maximum of the distribution}
  \item{m}{Columns of the output}
  \item{n}{Optional: Rows of the output matrix Default: 1}
}
\value{N            maximum}
\description{
 unidrnd(Lines,AnzData,1) generates random numbers for the discrete uniform distribution with maximum N.}
\author{
MT, reimplemented from matlabs  unidrnd}
