\name{About the statistics package}

\alias{statistics}

\docType{package}

\title{
A package containing useful functions to teach introductory statistics.
}

\description{
For discreet random variables, draws and calculates the probability of a certain number of
favourable outcomes out of a number of repetition of an experiment. For a continous random
variable, the graphics represents a normal curve with the area to the left or to the right
of a certain z or t value, or between two such values.
The package also contains functions to calculate the degrees of freedom and the pooled
standard deviation using the t distribution etc.
}

\seealso{dnorm, pnorm, dbinom}

\details{
    \tabular{ll}{
    Package: \tab statistics\cr
    Type: \tab Package\cr
    Version: \tab 0.8\cr
    Date: \tab 2024-10-29\cr
    License: \tab GPL-v3\cr
    }
}
\author{
Adrian Dusa

Maintainer: Adrian Dusa (dusa.adrian@unibuc.ro)
}

\keyword{misc}
